% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epipredict-data.R
\docType{data}
\name{can_prov_cases}
\alias{can_prov_cases}
\title{COVID Cases for Canada}
\format{
An object of class \code{epi_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 65299 rows and 4 columns.
}
\usage{
can_prov_cases
}
\description{
THIS IS A STUB This dataset covers the period from March 2020 to March 2023 and is
used in the [epipredict] slide vignette.
}
\keyword{datasets}
