% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epipredict-data.R
\docType{data}
\name{county_smoothed_cli_comparison}
\alias{county_smoothed_cli_comparison}
\title{Percent CLI from different surveys, compared to ground truth COVID incidence in a subset of counties}
\format{
An \code{\link[epiprocess:epi_df]{epiprocess::epi_df}} (object of class \code{c("epi_df", "tbl_df", "tbl", "data.frame")}) with 63840 rows and 5 columns.
}
\source{
This object contains modified data adapted from
...

The data is licensed under the terms of the
\href{url}{name}.

Modifications:
*
}
\usage{
county_smoothed_cli_comparison
}
\description{
Data set for more than 400 US counties containing CLI
(COVID-19-like-illness) incidence derived from two surveys, and a reference signal as
reported by JHU CSSE. This example data is a snapshot as of September 21,
2020, and ranges from April 11, 2020 to September 01, 2020.

The reference sign is based on reports made available
by the Center for Systems Science and Engineering at Johns Hopkins
University.

One survey is ... Google

One survey is ... CTIS in collaboration with Facebook.

Data is reported for counties that had at least 200 cumulative COVID-19 cases
on May 14, 2020, according to JHU CSSE.
}
\section{Data dictionary}{

The data has columns:
\describe{
\item{geo_value}{The 5-digit county FIPS code associated with each
row of measurements.}
\item{time_value}{The time value, a date in YYYY-MM-DD format}
\item{goog}{Seven-day average of CLI (covid-like-illness) cases}
\item{fb}{Seven-day average of CLI (covid-like-illness) cases}
\item{case}{Seven-day average of CLI (covid-like-illness) cases}
}
}

\keyword{datasets}
