% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epipredict-data.R
\docType{data}
\name{ctis_covid_behaviours}
\alias{ctis_covid_behaviours}
\title{CTIS COVID Behaviours}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1055 rows and 4 columns.
}
\source{
Data from Delphi COVIDcast. Obtained via the Delphi Epidata API.
\url{https://cmu-delphi.github.io/delphi-epidata/api/covidcast.html}. This
object contains a modified part of the
\href{https://cmu-delphi.github.io/delphi-epidata/symptom-survey/#covid-19-trends-and-impact-survey}{data
aggregations in the API} that are prepared from the
\href{https://www.pnas.org/doi/full/10.1073/pnas.2111454118}{COVID-19
Trends and Impact Survey}; see the first link for more information on
citing in publications. These aggregations are licensed under the terms of
the \href{https://creativecommons.org/licenses/by/4.0/}{Creative Commons
Attribution license}.

Modifications:
\itemize{
\item Selected a small subset of the available data.
\item Changed to wide format.
}

Delphi Group at Carnegie Mellon University
}
\usage{
ctis_covid_behaviours
}
\description{
Data set for a handful of states on masking and distancing behaviours
during the COVID-19 Pandemic and downloaded from the CMU Delphi COVIDcast
Epidata API. This data set covers the period from
June to December 2021.
}
\keyword{datasets}
