% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidatasets-package.R
\docType{package}
\name{epidatasets-package}
\alias{epidatasets}
\alias{epidatasets-package}
\title{epidatasets: Epidemiological Data for Delphi Tooling Examples}
\description{
This package contains data sets used to compile vignettes and other documentation in Delphi R Packages. The goal is to avoid calls to the Delphi Epidata API, and to deposit some examples here for easy offline use.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cmu-delphi/epidatasets}
  \item \url{https://cmu-delphi.github.io/epidatasets/}
}

}
\author{
\strong{Maintainer}: Nat DeFries \email{ndefries@andrew.cmu.edu}

Authors:
\itemize{
  \item Daniel J. McDonald \email{daniel@stat.ubc.ca}
}

Other contributors:
\itemize{
  \item Johns Hopkins University Center for Systems Science and Engineering (Owner of COVID-19 cases and deaths data from the COVID-19 Data Repository) [data contributor]
  \item Johns Hopkins University (Copyright holder of COVID-19 cases and deaths data from the COVID-19 Data Repository) [copyright holder]
  \item Carnegie Mellon University Delphi Group (Owner of masking, social-distancing, and CLI data from the COVID-19 Trends and Impacts Survey. Owner of claims-based CLI data from the Delphi Epidata API) [data contributor]
  \item The COVID-19 Canada Open Data Working Group (Owner of Canadian COVID-19 cases rates from the Covid19Canada data repository) [data contributor]
  \item Statistics Canada (Owner of Canadian graduate employment income data from the Statistics Canada website) [data contributor]
  \item Google (Collaborator on CLI data from the Google symptom surveys) [data contributor]
}

}
\keyword{internal}
