% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epipredict-data.R
\docType{data}
\name{ctis_covid_behaviours}
\alias{ctis_covid_behaviours}
\title{CTIS COVID Behaviours}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1055 rows and 4 columns.
}
\usage{
ctis_covid_behaviours
}
\description{
Data set for a handful of states on masking and distancing behaviours
during the COVID-19 Pandemic and downloaded from the CMU Delphi COVIDcast
Epidata API. This data set covers the period from
June to December 2021.
}
\keyword{datasets}
